backend/requirements.txt
fastapi
uvicorn
openai
langchain
sentence-transformers
faiss-cpu
pydantic
python-dotenv
sqlalchemy
joblib
scikit-learn
tqdm
pandas
python-dotenv
python-multipart
numpy
pandas
